+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+- Source ``.profile`` again when running psql admin commands on the
+  server.
+
 
 `1.0.20201004`_
 ~~~~~~~~~~~~~~~